Trait sp_settlement::SettlementApi

pub trait SettlementApi<Block: BlockT, DomainHash: Encode + Decode>: Core<Block> {

    // Provided methods
    fn execution_trace(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        domain_id: DomainId,
        receipt_hash: H256
    ) -> Result<Vec<DomainHash>, ApiError> { ... }
    fn execution_trace_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        domain_id: DomainId,
        receipt_hash: H256
    ) -> Result<Vec<DomainHash>, ApiError> { ... }
    fn state_root(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        domain_id: DomainId,
        domain_block_number: NumberFor<Block>,
        domain_block_hash: Block::Hash
    ) -> Result<Option<DomainHash>, ApiError> { ... }
    fn state_root_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        domain_id: DomainId,
        domain_block_number: NumberFor<Block>,
        domain_block_hash: Block::Hash
    ) -> Result<Option<DomainHash>, ApiError> { ... }
    fn primary_hash(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        domain_id: DomainId,
        domain_block_number: NumberFor<Block>
    ) -> Result<Option<Block::Hash>, ApiError> { ... }
    fn primary_hash_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        domain_id: DomainId,
        domain_block_number: NumberFor<Block>
    ) -> Result<Option<Block::Hash>, ApiError> { ... }
    fn receipts_pruning_depth(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn receipts_pruning_depth_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn head_receipt_number(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        domain_id: DomainId
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn head_receipt_number_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        domain_id: DomainId
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn oldest_receipt_number(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        domain_id: DomainId
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn oldest_receipt_number_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        domain_id: DomainId
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn maximum_receipt_drift(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn maximum_receipt_drift_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn extract_receipts(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        extrinsics: Vec<Block::Extrinsic>,
        domain_id: DomainId
    ) -> Result<Vec<ExecutionReceipt<NumberFor<Block>, Block::Hash, DomainHash>>, ApiError> { ... }
    fn extract_receipts_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        extrinsics: Vec<Block::Extrinsic>,
        domain_id: DomainId
    ) -> Result<Vec<ExecutionReceipt<NumberFor<Block>, Block::Hash, DomainHash>>, ApiError> { ... }
    fn extract_fraud_proofs(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        extrinsics: Vec<Block::Extrinsic>,
        domain_id: DomainId
    ) -> Result<Vec<FraudProof<NumberFor<Block>, Block::Hash>>, ApiError> { ... }
    fn extract_fraud_proofs_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        extrinsics: Vec<Block::Extrinsic>,
        domain_id: DomainId
    ) -> Result<Vec<FraudProof<NumberFor<Block>, Block::Hash>>, ApiError> { ... }
    fn submit_fraud_proof_unsigned(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        fraud_proof: FraudProof<NumberFor<Block>, Block::Hash>
    ) -> Result<(), ApiError> { ... }
    fn submit_fraud_proof_unsigned_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        fraud_proof: FraudProof<NumberFor<Block>, Block::Hash>
    ) -> Result<(), ApiError> { ... }

}

Provided Methods

fn execution_trace( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, domain_id: DomainId, receipt_hash: H256 ) -> Result<Vec<DomainHash>, ApiError>

Returns the trace of given domain receipt hash.

fn execution_trace_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, domain_id: DomainId, receipt_hash: H256 ) -> Result<Vec<DomainHash>, ApiError>

Returns the trace of given domain receipt hash.

fn state_root( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, domain_id: DomainId, domain_block_number: NumberFor<Block>, domain_block_hash: Block::Hash ) -> Result<Option<DomainHash>, ApiError>

Returns the state root of given domain block.

fn state_root_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, domain_id: DomainId, domain_block_number: NumberFor<Block>, domain_block_hash: Block::Hash ) -> Result<Option<DomainHash>, ApiError>

Returns the state root of given domain block.

fn primary_hash( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, domain_id: DomainId, domain_block_number: NumberFor<Block> ) -> Result<Option<Block::Hash>, ApiError>

Returns the primary block hash for given domain block number.

fn primary_hash_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, domain_id: DomainId, domain_block_number: NumberFor<Block> ) -> Result<Option<Block::Hash>, ApiError>

Returns the primary block hash for given domain block number.

fn receipts_pruning_depth( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<NumberFor<Block>, ApiError>

Returns the receipts pruning depth.

fn receipts_pruning_depth_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext ) -> Result<NumberFor<Block>, ApiError>

Returns the receipts pruning depth.

fn head_receipt_number( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, domain_id: DomainId ) -> Result<NumberFor<Block>, ApiError>

Returns the best execution chain number.

fn head_receipt_number_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, domain_id: DomainId ) -> Result<NumberFor<Block>, ApiError>

Returns the best execution chain number.

fn oldest_receipt_number( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, domain_id: DomainId ) -> Result<NumberFor<Block>, ApiError>

Returns the block number of oldest execution receipt.

fn oldest_receipt_number_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, domain_id: DomainId ) -> Result<NumberFor<Block>, ApiError>

Returns the block number of oldest execution receipt.

fn maximum_receipt_drift( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<NumberFor<Block>, ApiError>

Returns the maximum receipt drift.

fn maximum_receipt_drift_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext ) -> Result<NumberFor<Block>, ApiError>

Returns the maximum receipt drift.

fn extract_receipts( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, extrinsics: Vec<Block::Extrinsic>, domain_id: DomainId ) -> Result<Vec<ExecutionReceipt<NumberFor<Block>, Block::Hash, DomainHash>>, ApiError>

Extract the receipts from the given extrinsics.

fn extract_receipts_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, extrinsics: Vec<Block::Extrinsic>, domain_id: DomainId ) -> Result<Vec<ExecutionReceipt<NumberFor<Block>, Block::Hash, DomainHash>>, ApiError>

Extract the receipts from the given extrinsics.

fn extract_fraud_proofs( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, extrinsics: Vec<Block::Extrinsic>, domain_id: DomainId ) -> Result<Vec<FraudProof<NumberFor<Block>, Block::Hash>>, ApiError>

Extract the fraud proofs from the given extrinsics.

fn extract_fraud_proofs_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, extrinsics: Vec<Block::Extrinsic>, domain_id: DomainId ) -> Result<Vec<FraudProof<NumberFor<Block>, Block::Hash>>, ApiError>

Extract the fraud proofs from the given extrinsics.

fn submit_fraud_proof_unsigned( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, fraud_proof: FraudProof<NumberFor<Block>, Block::Hash> ) -> Result<(), ApiError>

Submits the fraud proof via an unsigned extrinsic.

fn submit_fraud_proof_unsigned_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, fraud_proof: FraudProof<NumberFor<Block>, Block::Hash> ) -> Result<(), ApiError>

Submits the fraud proof via an unsigned extrinsic.